+++ src/vue/modules/requests/update-asset-requests/components/request-actions.spec.js
@@ -31,7 +31,7 @@ describe('Update asset request actions', () => {
               accountId: 'SOME_ACCOUNT_ID',
             }),
           },
-        }
+        },
       },
     })